# 65x performance and memory optimisation for process citation