--- a/grobid-core/src/main/java/org/grobid/core/engines/AuthorParser.java
+++ b/grobid-core/src/main/java/org/grobid/core/engines/AuthorParser.java
@@ -1,5 +1,6 @@
 package org.grobid.core.engines;
 
+import java.util.regex.Pattern;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.collections4.CollectionUtils;
 
@@ -33,6 +34,7 @@ import java.util.StringTokenizer;
  * @author Patrice Lopez
  */
 public class AuthorParser {
+    public static final Pattern REGEX_PATTERN  = Pattern.compile("et\\.? al\\.?.*$");
 	private static Logger LOGGER = LoggerFactory.getLogger(AuthorParser.class);
     private final GenericTagger namesHeaderParser;
     private final GenericTagger namesCitationParser;
@@ -50,7 +52,7 @@ public class AuthorParser {
             return null;
         }
 
-        input = input.trim().replaceAll("et\\.? al\\.?.*$", " ");
+        input = REGEX_PATTERN.matcher(input.trim()).replaceAll(" ");
 
         // for language to English for the analyser to avoid any bad surprises
         List<LayoutToken> tokens = GrobidAnalyzer.getInstance().tokenizeWithLayoutToken(input, new Language("en", 1.0));
@@ -72,7 +74,7 @@ public class AuthorParser {
             return null;
         }
 
-        input = input.trim().replaceAll("et\\.? al\\.?.*$", " ");
+        input = REGEX_PATTERN.matcher(input.trim()).replaceAll(" ");
 
         // for language to English for the analyser to avoid any bad surprises
         List<LayoutToken> tokens = GrobidAnalyzer.getInstance().tokenizeWithLayoutToken(input, new Language("en", 1.0));