This PR intends to merge changes related to performance and memory optimisations
The details are posted below

1. pre-compile regex
   --> reduces mem usage
   --> improves performance as regex is precompiled
2. pre-init lexicon dependencies
   --> lazy initialization broke singleton in concurrent environment
   --> decreased memory footprint

We did GROBID performance profiling
Tools used:
Benchmarking tool: visualVM
Testing framework: locustio

Observations before optimisations:
![memory_before_optimisation](https://user-images.githubusercontent.com/65662582/88048831-a8aeea80-cb71-11ea-8ceb-7bdadde40cce.png)
Heap dump before optimisation. 
  
![threads_before_optimisation](https://user-images.githubusercontent.com/65662582/88048880-bcf2e780-cb71-11ea-9267-80653bbd68ea.png)
Thread dump before optimisation. 
  
We discovered following major issues around the code base in focus:
1. Regex Patterns were not precompiled
2. Lazy initialization of Singletons (can cause weak singleton)
3. String manipulations without StringBuilders

Observations after optimisations:
![memory_after_optimisation](https://user-images.githubusercontent.com/65662582/88048974-e3b11e00-cb71-11ea-9438-54359d2e634a.png)
Heap dump after optimisation. 
  

![threads_after_optimisation](https://user-images.githubusercontent.com/65662582/88049038-02171980-cb72-11ea-9480-891d4d6a96d0.png)
Thread dump after optimisations. 
  



  | Before optimization | After optimization
-- | -- | --
GROBID performance | <ul><li>Total users: 100.</li><li>Total runtime: 600s (10 min)</li><li> Total requests: 123<li>Failures: 121</li><li> Throughput: 0.21 requests/sec</li><li> Memory consumption 4 GB </li></ul> | <ul><li> Total users: 100</li><li> Total runtime: 600s (10 min)</li><li> Total requests: 7915</li><li> Failures: 0</li><li> Throughput: 13.17 requests/sec</li><li> Memory consumption 750 MB</li></ul>



Co-authored-by: rvkhakhkhar <rvkhakhkhar@gmail.com>
Co-authored-by: imprashant <imprashantm@gmail.com>