@kermitt2 2020-07-22
Thanks a lot for the PR and the effort to improve the processing of citations!

I did some performance tests on my side with the current master and this PR version, but did not see differences. Using jmeter and jprofiler on 100,000 processCitation queries with 100 concurrent queries, I have the following with the PR:

![grobid-pr607-summary](https://user-images.githubusercontent.com/2340795/88133535-9de06e00-cbe2-11ea-8158-d9e98caeb00e.png)
![grobid-pr607-jprofiler](https://user-images.githubusercontent.com/2340795/88133542-a173f500-cbe2-11ea-9768-42f64ad638bc.png)

And with the master version:

![grobid-master-summary](https://user-images.githubusercontent.com/2340795/88133559-adf84d80-cbe2-11ea-8ad4-32fe641c5b4c.png)
![grobid-master-jprofiler](https://user-images.githubusercontent.com/2340795/88133568-b2246b00-cbe2-11ea-81e0-bddca955b9c5.png)

So I don't see any significant difference, both for memory and runtime, which is more or less what I would expect because:
- the use of regular expressions is really negligible in comparison to the full chain of ML processing, so optimizing them could not really impact the overall runtime 
- lexicon init is done one time

These changes could not explain a difference of 65 times I think, because they touch very minor aspects of the whole process. 

So far I've not seen issues with the lazy initialization of the lexicon, the benchmarking that I am doing regularly gives the same results with one or several threads. Luca fixed the problem of the lazy init for the resources of the language recognizer a few month ago with PR #585, but I don't think it is similar for the full lexicon object. However I can change the lazy initialization of the lexicon in a similar manner (double check synchro on the instance) to be sure.

Now how to explain the 65 times difference you report? I could imagine one of the following:
- you ran the master version with consolidation option, while the PR is running without
- the evaluation of the master version was done on the server "cold". The loading of the models and lexical information takes some time and the server might have been overloaded with queries at the same time, resulting in a lot of 503 responses and some delay. But after a while, it should be on the correct track, and I doubt it could explain a 65x difference alone.
- you are using a Windows environment where the multithreading of the current master appears to be broken (we are focusing on Linux and macos support so far, given our limited availability for this project)