@khakhkharr 2020-07-22
@kermitt2 

Thanks for responding to our PR.
Just to correct, we using MacOS environment for performance testing.

We will get get back to you with more accurate information about what data, script and configuraion we used to benchmark `api/processCitations` without consolidation option.