@kermitt2 2021-03-02
I didn't close it as a reminder to add the lazy init of Lexicon (but it is added in PR #687 with other changes on Lexicon.java) and to pre-compile the regex (not done yet.... though I did not observe actual runtime change, it would be good to add). 

In PR #687 there is a a runtime gain of around 10%, but it was by improving the alignment between document layout tokens and labelled tokens returned by the sequence labelling algorithm, which is something I spotted by the breakdown of runtime per method produced by JProfiler. The total time spent on these regex is really small (less than 2% for full document processing). 

One interesting "hotspot" for runtime is however the time spent by FastMatcher (around 15% of the whole runtime), in particular matching the journal names (11.3%). There are apparently too many short abbreviated concurrent journal names matching at the same time in text body content, and this could certainly be optimized.