@kermitt2 2021-12-19
Issue about potential optimization for FastMatcher and journal name: #731

Pre-compiled regex added in f466930dbca9ab39b4f9a50b1df260a973b6cab4 which adds a few percents speed-up following some quick tests. 

Thank you again @murkutep for the issue !